% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{udp_callback_logger}
\alias{udp_callback_logger}
\title{Log UDP broadcast packets to a file}
\usage{
udp_callback_logger(port, callback)
}
\arguments{
\item{port}{UDP broadcast port to listen on}

\item{callback}{R callback function (see Details)}
}
\value{
none
}
\description{
Listen for broadcast payloads on \code{port} and call \code{callback} with the payload data.
The listener will continue forever or until a \code{CTRL-C} (\code{SIGINT}) is encountered.
}
\details{
The \code{callback} function must take one parameter which contains the payload data.
The callback return value should be \code{TRUE} if the listener should be stopped and
\code{FALSE} if it should continue listening for broadcast payloads and calling the
\code{callback} function.
}
\note{
The current implementation of the logger assumes character payload data.
Please file an issue if you need it to handle raw vectors.
}
\examples{
\dontrun{
udp_callback_logger(22222L, \(x) { writeLines(x); return(FALSE) })
}
}
