% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{udp_file_logger}
\alias{udp_file_logger}
\title{Log UDP broadcast packets to a file}
\usage{
udp_file_logger(port, path)
}
\arguments{
\item{port}{UDP broadcast port to listen on}

\item{path}{path (including filename) to write records to}
}
\value{
none
}
\description{
Listen for broadcast payloads on \code{port} and log them directly to \code{file}.
The listener will continue forever or until a \code{CTRL-C} (\code{SIGINT}) is encountered.
}
\note{
The current implementation of the logger assumes character payload data.
Please file an issue if you need it to handle raw vectors.
}
\examples{
\dontrun{
udp_file_logger(22222L, "~/Data/davis.json")
}
}
