% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpbroadcastreceiver-package.R
\docType{package}
\name{udpbroadcastreceiver}
\alias{udpbroadcastreceiver}
\alias{udpbroadcastreceiver-package}
\title{Receive and Log or Process UDP Broadcast Payloads}
\description{
Many remote sensors, such as weather stations, have the ability to broadcast
their record payloads over UDP. Tools are provided to listen for these broadcast packets
and either log them directly to a file or process them with callback functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/udpbroadcastreceiver}
  \item Report bugs at \url{https://github.com/hrbrmstr/udpbroadcastreceiver/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
