% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpbroadcastreceiver-package.R
\docType{package}
\name{udpbroadcastreceiver}
\alias{udpbroadcastreceiver}
\alias{udpbroadcastreceiver-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/udpbroadcastreceiver}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/udpbroadcastreceiver/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
